fix up recursion report and report each file as its being parsed

+++ parser/cpp/Parser.cpp
@@ -346,7 +346,8 @@ std::vector< ParseState > ParseRecursive(
 		{
 			if( ParseName( axTokens,
                 axNames, xGrammar,
-                i, axProductions[ i ],
+                static_cast< int >( i ),
+                axProductions[ i ],
                 iCurrentListCount, j,
 				axWorkingNewStates, axNewStates ) )
 			{
